+++ imagelab_react/package.json
@@ -18,6 +18,7 @@
     "react": "^18.2.0",
     "react-blockly": "^7.2.2",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "react-zoom-pan-pinch": "^3.1.0",
     "sharp": "^0.32.1",
