+++ imagelab_react/src/components/Dashboard/Dashboard.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Dashboard = () => {
+  return (
+    <div>
+      Dashboard
+    </div>
+  )
+}
+
+export default Dashboard
