+++ imagelab_react/src/components/Profile/Login.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Login = () => {
+  return (
+    <div>
+      Login
+    </div>
+  )
+}
+
+export default Login
