+++ imagelab_react/src/components/Profile/Profile.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Profile = () => {
+  return (
+    <div>
+      Profile
+    </div>
+  )
+}
+
+export default Profile
