+++ imagelab_react/src/components/Profile/Signup.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Signup = () => {
+  return (
+    <div>
+      Signup
+    </div>
+  )
+}
+
+export default Signup
