+++ imagelab_react/src/components/Workspace/browse-blocks-functions.js
