+++ imagelab_react/src/components/Workspace/imagelab-block.js
