+++ imagelab_react/src/components/Workspace/toolboxConfiguration.js
