+++ imagelab_react/src/components/Workspace/workspaceConfig.js
